fix requite-cycle warning
